+++ pom.xml
@@ -3,7 +3,7 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>com.lambazon</groupId>
+    <groupId>com.parkit</groupId>
     <artifactId>parking-system</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>jar</packaging>
@@ -76,7 +76,7 @@
                             <archive>
                                 <manifest>
                                     <mainClass>
-                                        com.lambazon.parkingsystem.App
+                                        com.parkit.parkingsystem.App
                                     </mainClass>
                                 </manifest>
                             </archive>
