--- src/main/java/com/lambazon/parkingsystem/App.java
@@ -1,13 +1,0 @@
-package com.lambazon.parkingsystem;
-
-import com.lambazon.parkingsystem.service.InteractiveShell;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-public class App {
-    private static final Logger logger = LogManager.getLogger("App");
-    public static void main(String args[]){
-        logger.info("Initializing Parking System");
-        InteractiveShell.loadInterface();
-    }
-}
