--- src/main/java/com/lambazon/parkingsystem/constants/Fare.java
@@ -1,6 +1,0 @@
-package com.lambazon.parkingsystem.constants;
-
-public class Fare {
-    public static final double BIKE_RATE_PER_HOUR = 1.0;
-    public static final double CAR_RATE_PER_HOUR = 1.5;
-}
