--- src/main/java/com/lambazon/parkingsystem/constants/ParkingType.java
@@ -1,6 +1,0 @@
-package com.lambazon.parkingsystem.constants;
-
-public enum ParkingType {
-    CAR,
-    BIKE
-}
