+++ src/main/java/com/parkit/parkingsystem/constants/Fare.java
@@ -1,0 +1,6 @@
+package com.parkit.parkingsystem.constants;
+
+public class Fare {
+    public static final double BIKE_RATE_PER_HOUR = 1.0;
+    public static final double CAR_RATE_PER_HOUR = 1.5;
+}
