+++ src/main/java/com/parkit/parkingsystem/constants/ParkingType.java
@@ -1,0 +1,6 @@
+package com.parkit.parkingsystem.constants;
+
+public enum ParkingType {
+    CAR,
+    BIKE
+}
