+++ src/test/java/com/parkit/parkingsystem/integration/service/DataBasePrepareService.java
@@ -1,0 +1,30 @@
+package com.parkit.parkingsystem.integration.service;
+
+import com.parkit.parkingsystem.integration.config.DataBaseTestConfig;
+
+import java.sql.Connection;
+
+public class DataBasePrepareService {
+
+    DataBaseTestConfig dataBaseTestConfig = new DataBaseTestConfig();
+
+    public void clearDataBaseEntries(){
+        Connection connection = null;
+        try{
+            connection = dataBaseTestConfig.getConnection();
+
+            //set parking entries to available
+            connection.prepareStatement("update parking set available = true").execute();
+
+            //clear ticket entries;
+            connection.prepareStatement("truncate table ticket").execute();
+
+        }catch(Exception e){
+            e.printStackTrace();
+        }finally {
+            dataBaseTestConfig.closeConnection(connection);
+        }
+    }
+
+
+}
